pyproject.toml
[project]
name = "stock-trader-o3-algo"
version = "0.1.0"
description = "Algorithmic trading strategy based on the micro-CTA approach"
authors = [
    {name = "User"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas",
    "numpy",
    "yfinance",
    "alpaca-trade-api",
    "pytz",
    "matplotlib",
    "pytest",
    "python-dotenv"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
